Fix bug in opendir, invalid param handler for VS8

+++ win32/ntport.h
@@ -307,7 +307,7 @@ extern void dprintf(char *,...);
 /* support.c */
 extern void nt_init(void);
 extern int gethostname(char*,int);
-extern char* forward_slash_get_cwd(char *,int len );
+extern char* forward_slash_get_cwd(char *,size_t len );
 extern int  nt_chdir(char*);
 extern void  nt_execve(char *,char**,char**);
 extern void  nt_exec(char *,char**);
